requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
motor
pymongo
python-dotenv
pydantic
apscheduler
